requirements.txt
# Core dependencies
numpy>=1.24.0
pandas>=2.0.0
scipy>=1.10.0
scikit-learn>=1.2.0
pytest>=8.3.5
coverage>=7.4.0
psutil>=5.9.8
pyyaml>=6.0.1
python-jwt>=4.0.0
bcrypt>=4.1.2
watchdog>=3.0.0
requests>=2.31.0

# Trading and exchange
pybit>=2.4.1
python-binance>=1.0.16
ccxt>=3.0.0

# Data processing and analysis
ta-lib>=0.4.24
python-binance>=1.0.16
vectorbt>=0.24.0

# Machine Learning
tensorflow>=2.12.0
torch>=2.0.0
transformers>=4.28.0
statsmodels>=0.14.0

# Database and storage
SQLAlchemy>=2.0.0
aiosqlite>=0.19.0

# API and networking
aiohttp>=3.8.4
requests>=2.28.0
websockets>=11.0.0

# Configuration and environment
python-dotenv>=1.0.0
PyYAML>=6.0.0

# System monitoring
psutil>=5.9.0

# Utilities
python-dateutil>=2.8.2
pytz>=2023.3
tqdm>=4.65.0

# Testing
pytest>=7.3.1
pytest-asyncio>=0.21.0
pytest-cov>=4.0.0

# Development tools
black>=23.3.0
flake8>=6.0.0
mypy>=1.2.0
pylint>=2.17.0

# Security
cryptography>=42.0.0
python-jose>=3.3.0

# Testing
pytest-asyncio>=0.23.5
pytest-cov>=4.1.0
pytest-xdist>=3.5.0

# Monitoring
prometheus_client>=0.19.0
statsd>=4.0.1

# Logging
python-json-logger>=2.0.7
structlog>=24.1.0

# Database
sqlalchemy>=2.0.25
alembic>=1.13.1

# API
fastapi>=0.109.0
uvicorn>=0.27.0
python-multipart>=0.0.6

# Utils
python-dateutil>=2.8.2
pytz>=2024.1
